@@ -4,7 +4,7 @@
   dependencies = ["parver>=0.2.1", "pulumi>=3.91.1,<4.0.0", "pulumi-aws>=6.0.4,<7.0.0", "pulumi-docker>=4.5.1,<5.0.0", "semver>=2.8.1"]
   keywords = ["pulumi", "aws", "awsx", "kind/component", "category/cloud"]
   readme = "README.md"
-  requires-python = ">=3.7"
+  requires-python = ">=3.8"
   version = "0.0.0"
   [project.license]
     text = "Apache-2.0"
