@@ -8,7 +8,6 @@ import (
 	"reflect"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-	"github.com/pulumi/pulumi/sdk/v3/go/pulumix"
 )
 
 type TaskDefinitionPortMappingAppProtocol string
@@ -79,12 +78,6 @@ func (o TaskDefinitionPortMappingAppProtocolOutput) ToTaskDefinitionPortMappingA
 	}).(TaskDefinitionPortMappingAppProtocolPtrOutput)
 }
 
-func (o TaskDefinitionPortMappingAppProtocolOutput) ToOutput(ctx context.Context) pulumix.Output[TaskDefinitionPortMappingAppProtocol] {
-	return pulumix.Output[TaskDefinitionPortMappingAppProtocol]{
-		OutputState: o.OutputState,
-	}
-}
-
 func (o TaskDefinitionPortMappingAppProtocolOutput) ToStringOutput() pulumi.StringOutput {
 	return o.ToStringOutputWithContext(context.Background())
 }
@@ -120,12 +113,6 @@ func (o TaskDefinitionPortMappingAppProtocolPtrOutput) ToTaskDefinitionPortMappi
 	return o
 }
 
-func (o TaskDefinitionPortMappingAppProtocolPtrOutput) ToOutput(ctx context.Context) pulumix.Output[*TaskDefinitionPortMappingAppProtocol] {
-	return pulumix.Output[*TaskDefinitionPortMappingAppProtocol]{
-		OutputState: o.OutputState,
-	}
-}
-
 func (o TaskDefinitionPortMappingAppProtocolPtrOutput) Elem() TaskDefinitionPortMappingAppProtocolOutput {
 	return o.ApplyT(func(v *TaskDefinitionPortMappingAppProtocol) TaskDefinitionPortMappingAppProtocol {
 		if v != nil {
@@ -150,10 +137,12 @@ func (o TaskDefinitionPortMappingAppProtocolPtrOutput) ToStringPtrOutputWithCont
 	}).(pulumi.StringPtrOutput)
 }
 
-// TaskDefinitionPortMappingAppProtocolInput is an input type that accepts TaskDefinitionPortMappingAppProtocolArgs and TaskDefinitionPortMappingAppProtocolOutput values.
-// You can construct a concrete instance of `TaskDefinitionPortMappingAppProtocolInput` via:
+// TaskDefinitionPortMappingAppProtocolInput is an input type that accepts values of the TaskDefinitionPortMappingAppProtocol enum
+// A concrete instance of `TaskDefinitionPortMappingAppProtocolInput` can be one of the following:
 //
-//	TaskDefinitionPortMappingAppProtocolArgs{...}
+//	TaskDefinitionPortMappingAppProtocolHttp
+//	TaskDefinitionPortMappingAppProtocolHttp2
+//	TaskDefinitionPortMappingAppProtocolGrpc
 type TaskDefinitionPortMappingAppProtocolInput interface {
 	pulumi.Input
 
@@ -186,12 +175,6 @@ func (in *taskDefinitionPortMappingAppProtocolPtr) ToTaskDefinitionPortMappingAp
 
 func (in *taskDefinitionPortMappingAppProtocolPtr) ToTaskDefinitionPortMappingAppProtocolPtrOutputWithContext(ctx context.Context) TaskDefinitionPortMappingAppProtocolPtrOutput {
 	return pulumi.ToOutputWithContext(ctx, in).(TaskDefinitionPortMappingAppProtocolPtrOutput)
-}
-
-func (in *taskDefinitionPortMappingAppProtocolPtr) ToOutput(ctx context.Context) pulumix.Output[*TaskDefinitionPortMappingAppProtocol] {
-	return pulumix.Output[*TaskDefinitionPortMappingAppProtocol]{
-		OutputState: in.ToTaskDefinitionPortMappingAppProtocolPtrOutputWithContext(ctx).OutputState,
-	}
 }
 
 func init() {
